+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/TextAnalyticsActions.cs
@@ -69,5 +69,14 @@ namespace Azure.AI.TextAnalytics
         /// This property only applies for <see cref="TextAnalyticsClientOptions.ServiceVersion.V3_2_Preview_2"/> and up.
         /// </remarks>
         public IReadOnlyCollection<SingleCategoryClassifyAction> SingleCategoryClassifyActions { get; set; }
+
+        /// <summary>
+        /// The set of <see cref="MultiCategoryClassifyAction"/> that will get executed on the input documents.
+        /// Note that currently only one <see cref="MultiCategoryClassifyAction"/> is supported.
+        /// </summary>
+        /// <remarks>
+        /// This property only applies for <see cref="TextAnalyticsClientOptions.ServiceVersion.V3_2_Preview_2"/> and up.
+        /// </remarks>
+        public IReadOnlyCollection<MultiCategoryClassifyAction> MultiCategoryClassifyActions { get; set; }
     }
 }