+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/TextAnalyticsClient.cs
@@ -2478,6 +2478,10 @@ namespace Azure.AI.TextAnalytics
             {
                 tasks.CustomSingleClassificationTasks = Transforms.ConvertFromSingleCategoryClassifyActionsToTasks(actions.SingleCategoryClassifyActions);
             }
+            if (actions.MultiCategoryClassifyActions != null)
+            {
+                tasks.CustomMultiClassificationTasks = Transforms.ConvertFromMultiCategoryClassifyActionsToTasks(actions.MultiCategoryClassifyActions);
+            }
             return tasks;
         }
 
@@ -2489,7 +2493,9 @@ namespace Azure.AI.TextAnalytics
                 actions.ExtractKeyPhrasesActions?.Count > 1 ||
                 actions.AnalyzeSentimentActions?.Count > 1 ||
                 actions.ExtractSummaryActions?.Count > 1 ||
-                actions.SingleCategoryClassifyActions?.Count > 1)
+                actions.SingleCategoryClassifyActions?.Count > 1 ||
+                actions.MultiCategoryClassifyActions?.Count > 1
+                )
             {
                 throw new ArgumentException("Multiple of the same action is not currently supported.");
             }