+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/Infrastructure/TextAnalyticsTestEnvironment.cs
@@ -11,5 +11,7 @@ namespace Azure.AI.TextAnalytics.Tests
         public string ApiKey => GetRecordedVariable("TEXT_ANALYTICS_API_KEY", options => options.IsSecret());
         public string SingleClassificationProjectName = "659c1851-be0b-4142-b12a-087da9785926";
         public string SingleClassificationDeploymentName = "659c1851-be0b-4142-b12a-087da9785926";
+        public string MultiClassificationProjectName = "7cdace98-537b-494a-b69a-c19754718025";
+        public string MultiClassificationDeploymentName = "7cdace98-537b-494a-b69a-c19754718025";
     }
 }