+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/TextAnalyticsModelFactoryTests.cs
@@ -142,6 +142,12 @@ namespace Azure.AI.TextAnalytics.Tests
                 TextAnalyticsModelFactory.SingleCategoryClassifyActionResult(default, default)
             };
 
+            var multiCategoryClassifyActionResults = new List<MultiCategoryClassifyActionResult>()
+            {
+                TextAnalyticsModelFactory.MultiCategoryClassifyActionResult(default, default),
+                TextAnalyticsModelFactory.MultiCategoryClassifyActionResult(default, default)
+            };
+
             var actionsResult = TextAnalyticsModelFactory.AnalyzeActionsResult(
                 extractKeyPhrasesActionResults,
                 recognizeEntitiesActionResults,
@@ -169,7 +175,8 @@ namespace Azure.AI.TextAnalytics.Tests
                 recognizeLinkedEntitiesActionResults,
                 analyzeSentimentActionResults,
                 extractSummaryActionResults,
-                singleCategoryClassifyActionResults);
+                singleCategoryClassifyActionResults,
+                multiCategoryClassifyActionResults);
 
             CollectionAssert.AreEquivalent(extractKeyPhrasesActionResults, actionsResult.ExtractKeyPhrasesResults);
             CollectionAssert.AreEquivalent(recognizeEntitiesActionResults, actionsResult.RecognizeEntitiesResults);
@@ -178,6 +185,7 @@ namespace Azure.AI.TextAnalytics.Tests
             CollectionAssert.AreEquivalent(analyzeSentimentActionResults, actionsResult.AnalyzeSentimentResults);
             CollectionAssert.AreEquivalent(extractSummaryActionResults, actionsResult.ExtractSummaryResults);
             CollectionAssert.AreEquivalent(singleCategoryClassifyActionResults, actionsResult.SingleCategoryClassifyResults);
+            CollectionAssert.AreEquivalent(multiCategoryClassifyActionResults, actionsResult.MultiCategoryClassifyResults);
         }
 
         [Test]