+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/samples/Sample11_MultiCategoryClassifyConvenienceAsync.cs
@@ -13,15 +13,15 @@ namespace Azure.AI.TextAnalytics.Samples
     public partial class TextAnalyticsSamples : SamplesBase<TextAnalyticsTestEnvironment>
     {
         [Test]
-        public async Task SingleCategoryClassifyConvenienceAsync()
+        public async Task MultiCategoryClassifyConvenienceAsync()
         {
-            // Create a Text Analytics client.
+            // Create a text analytics client.
             string endpoint = TestEnvironment.Endpoint;
             string apiKey = TestEnvironment.ApiKey;
 
             var client = new TextAnalyticsClient(new Uri(endpoint), new AzureKeyCredential(apiKey));
 
-            #region Snippet:TextAnalyticsSingleCategoryClassifyAsync
+            #region Snippet:TextAnalyticsMultiCategoryClassifyAsync
             // Get input document.
             string document = @"I need a reservation for an indoor restaurant in China. Please don't stop the music. Play music and add it to my playlist.";
 
@@ -37,24 +37,24 @@ namespace Azure.AI.TextAnalytics.Samples
             string projectName = "<projectName>";
             string deploymentName = "<deploymentName>";
 #else
-            string projectName = TestEnvironment.SingleClassificationProjectName;
-            string deploymentName = TestEnvironment.SingleClassificationDeploymentName;
+            string projectName = TestEnvironment.MultiClassificationProjectName;
+            string deploymentName = TestEnvironment.MultiClassificationDeploymentName;
 #endif
 
-            var singleCategoryClassifyAction = new SingleCategoryClassifyAction(projectName, deploymentName);
+            var multiCategoryClassifyAction = new MultiCategoryClassifyAction(projectName, deploymentName);
 
             TextAnalyticsActions actions = new TextAnalyticsActions()
             {
-                SingleCategoryClassifyActions = new List<SingleCategoryClassifyAction>() { singleCategoryClassifyAction }
+                MultiCategoryClassifyActions = new List<MultiCategoryClassifyAction>() { multiCategoryClassifyAction }
             };
 
             // Start analysis process.
             AnalyzeActionsOperation operation = await client.StartAnalyzeActionsAsync(batchInput, actions);
 
             await operation.WaitForCompletionAsync();
-            #endregion Snippet:TextAnalyticsSingleCategoryClassifyAsync
+            #endregion Snippet:TextAnalyticsMultiCategoryClassifyAsync
 
-            #region Snippet:TextAnalyticsSingleCategoryClassifyOperationStatus
+            #region Snippet:TextAnalyticsMultiCategoryClassifyOperationStatus
             // View operation status.
             Console.WriteLine($"AnalyzeActions operation has completed");
             Console.WriteLine();
@@ -65,24 +65,33 @@ namespace Azure.AI.TextAnalytics.Samples
             Console.WriteLine($"Status       : {operation.Status}");
             Console.WriteLine($"Last Modified: {operation.LastModified}");
             Console.WriteLine();
-            #endregion Snippet:TextAnalyticsSingleCategoryClassifyOperationStatus
+            #endregion Snippet:TextAnalyticsMultiCategoryClassifyOperationStatus
 
-            #region Snippet:TextAnalyticsSingleCategoryClassifyAsyncViewResults
+            #region Snippet:TextAnalyticsMultiCategoryClassifyAsyncViewResults
             // View operation results.
             await foreach (AnalyzeActionsResult documentsInPage in operation.Value)
             {
-                IReadOnlyCollection<SingleCategoryClassifyActionResult> singleClassificationActionResults = documentsInPage.SingleCategoryClassifyResults;
+                IReadOnlyCollection<MultiCategoryClassifyActionResult> multiClassificationActionResults = documentsInPage.MultiCategoryClassifyResults;
 
-                foreach (SingleCategoryClassifyActionResult classificationActionResults in singleClassificationActionResults)
+                foreach (MultiCategoryClassifyActionResult classificationActionResults in multiClassificationActionResults)
                 {
-                    foreach (SingleCategoryClassifyResult documentResults in classificationActionResults.DocumentsResults)
+                    foreach (MultiCategoryClassifyResult documentResults in classificationActionResults.DocumentsResults)
                     {
-                        Console.WriteLine($"  Class category \"{documentResults.Classification.Category}\" predicted with a confidence score of {documentResults.Classification.ConfidenceScore}.");
-                        Console.WriteLine();
+                        if (documentResults.Classifications.Count > 0)
+                        {
+                            Console.WriteLine($"  The following classes were predicted for this document:");
+
+                            foreach (ClassificationCategory classification in documentResults.Classifications)
+                            {
+                                Console.WriteLine($"  Class category \"{classification.Category}\" predicted with a confidence score of {classification.ConfidenceScore}.");
+                            }
+
+                            Console.WriteLine();
+                        }
                     }
                 }
             }
-            #endregion Snippet:TextAnalyticsSingleCategoryClassifyAsyncViewResults
+            #endregion Snippet:TextAnalyticsMultiCategoryClassifyAsyncViewResults
         }
     }
 }